[Android] Changes for partner customizations
Chromium change:
https://github.com/chromium/chromium/commit/254d31968ece2b5bb58984ed2f79ae01d001c9d1

Remove partnercustomizations AppHooks dep and modularize code
* Remove AppHooks dependency for partnercustomizations.
* Move partnercustomizations code to its own module.
* Move test rules to the partnercustomizations module and setup BUILD.
* Move tests to the partnercustomizations module.

Bug: 1291657
