+++ android/java/apk_for_test.flags
@@ -453,6 +453,14 @@
     *** get(...);
 }
 
+-keep class org.chromium.chrome.browser.partnercustomizations.CustomizationProviderDelegateImpl {
+    public <init>(...);
+}
+
+-keep class org.chromium.chrome.browser.partnercustomizations.BraveCustomizationProviderDelegateImpl {
+    public <init>(...);
+}
+
 -keep class org.chromium.chrome.browser.share.send_tab_to_self.DevicePickerBottomSheetContent {
     public <init>(...);
     *** createManageDevicesLink(...);
