+++ browser/partnercustomizations/BUILD.gn
@@ -1,0 +1,10 @@
+# Copyright 2022 The Brave Authors. All rights reserved.
+# Use of this source code is governed by a BSD-style license that can be
+# found in the LICENSE file.
+
+import("//build/config/android/rules.gni")
+
+android_library("delegate_public_impl_java") {
+  sources = [ "java/src/org/chromium/chrome/browser/partnercustomizations/BraveCustomizationProviderDelegateImpl.java" ]
+  deps = [ "//chrome/browser/partnercustomizations:delegate_java" ]
+}
